Enemy Behaviour
//#. Change
Import Cinemachine
Enemy Interaction: Scenes zoom in on the same scene, not scene transition.
Door Interaction: Movement in the same scene, not scene transition.
※It is necessary to simplify the code more.
For example, how to find players instead of public
However, because the game is a burden, it is better to refrain from using the 'Find' function to find the player.

//#.EnemyAILogic
RoamingAI: AI roaming without falling when it's not in a combat state.

//#.BattleLogic
When a player enters a combat state, zoom in and focus between the enemy and the player.
The Player must approach the enemy to see HP.

Only when the player is in combat with the enemy does the damage and take damage.
It also damages with a random value. (Damage adjustments can be made later)
※It would be better to change the system to an event trigger instead of a button. If the player raises the mouse to the enemy, it looks like it can attack even if it's not currently in combat. (it's impossible to attack)